Service/requirements.txt
torch==1.11.0
torchvision==0.12.0
timm==0.5.4
numpy==1.22.3
Pillow==9.1.0
Flask==2.1.2
PyYAML==6.0
easydict